+++ public/local/kernel.js
@@ -6306,7 +6306,7 @@ var require_config = __commonJS({
       return import_buffer.Buffer.from(hex, "hex");
     }
     var config = {
-      version: true ? "1.9.0" : "local",
+      version: true ? "1.9.1" : "local",
       env: "local",
       isProd: true,
       isTest: false,